Convert JavaScriptCore to Modern Objective-C with ARC https://bugs.webkit.org/show_bug.cgi?id=251797
Reviewed by NOBODY (OOPS!).

This was a while in the making. This focuses on modernizing
and migrating the JavaScriptCore code to ARC, which will
require an API review, but nonetheless, with the code
upgraded and running under ARC,this is one incremental step
in the migration to ARC done.

* Source/JavaScriptCore/API/APICallbackFunction.h:
* Source/JavaScriptCore/API/JSAPIGlobalObject.mm:
* Source/JavaScriptCore/API/JSAPIWrapperObject.mm:
* Source/JavaScriptCore/API/JSContext.mm:
* Source/JavaScriptCore/API/JSContextInternal.h:
* Source/JavaScriptCore/API/JSManagedValue.h:
* Source/JavaScriptCore/API/JSManagedValue.mm:
* Source/JavaScriptCore/API/JSScript.h:
* Source/JavaScriptCore/API/JSScript.mm:
* Source/JavaScriptCore/API/JSScriptInternal.h:
* Source/JavaScriptCore/API/JSScriptSourceProvider.mm:
* Source/JavaScriptCore/API/JSValue.mm:
* Source/JavaScriptCore/API/JSVirtualMachine.mm:
* Source/JavaScriptCore/API/JSWrapperMap.mm:
* Source/JavaScriptCore/API/ObjCCallbackFunction.mm:
* Source/JavaScriptCore/API/tests/CurrentThisInsideBlockGetterTest.mm:
* Source/JavaScriptCore/API/tests/DateTests.mm:
* Source/JavaScriptCore/API/tests/Regress141275.mm:
* Source/JavaScriptCore/API/tests/Regress141809.mm:
* Source/JavaScriptCore/API/tests/testapi.mm:
* Source/JavaScriptCore/JavaScriptCore.xcodeproj/project.pbxproj:
* Source/JavaScriptCore/inspector/ContentSearchUtilities.cpp:
* Source/JavaScriptCore/inspector/remote/cocoa/RemoteConnectionToTargetCocoa.mm:
* Source/JavaScriptCore/inspector/remote/cocoa/RemoteInspectorCocoa.mm:
* Source/JavaScriptCore/inspector/remote/cocoa/RemoteInspectorXPCConnection.mm:
* Source/JavaScriptCore/inspector/scripts/codegen/objc_generator.py:
* Source/JavaScriptCore/inspector/scripts/codegen/objc_generator_templates.py:
* Source/JavaScriptCore/inspector/scripts/tests/expected/type-declaration-object-type.json-result:
* Source/JavaScriptCore/inspector/scripts/tests/expected/type-requiring-runtime-casts.json-result:
* Source/JavaScriptCore/testmem/testmem.mm:

+++ Source/JavaScriptCore/dfg/DFGScoreBoard.h
@@ -138,22 +138,22 @@ public:
     void dump()
     {
         dataLogF("    USED: [ ");
-        for (unsigned i = 0; i < m_used.size(); ++i) {
+        for (size_t i = 0; i < m_used.size(); ++i) {
             if (!m_free.contains(i)) {
-                dataLogF("%d:", i);
+                dataLogF("%zu:", i);
                 if (m_used[i] == max())
                     dataLogF("local ");
                 else
-                    dataLogF("%d ", m_used[i]);
+                    dataLogF("%u ", m_used[i]);
             }
         }
         dataLogF("]\n");
 
         dataLogF("    FREE: [ ");
-        for (unsigned i = 0; i < m_used.size(); ++i) {
+        for (size_t i = 0; i < m_used.size(); ++i) {
             if (m_free.contains(i) && m_used[i] != max()) {
                 ASSERT(!m_used[i]);
-                dataLogF("%d ", i);
+                dataLogF("%zu ", i);
             }
         }
         dataLogF("]\n");
